+++ CHANGELOG.md
@@ -7,8 +7,8 @@ _To ensure you're using the latest version, open the game and press <kbd>Ctrl</k
 ## [v1.5.0](https://github.com/arcomage/arcomage-hd/compare/v1.4.0...v1.5.0) (2025-?-?) (in progress)
 
 - Full control with mouse, touch device, keyboard, or gamepad 🖱️📱⌨️🎮 ([instructions](gamecontrols.md))
-- [Show reasons for win/loss/draw](https://github.com/arcomage/arcomage-hd/issues/84)
-- [Optimize AnimatedNumber](https://github.com/arcomage/arcomage-hd/issues/95) & lang change
+- At game end screen, [show reasons for win/loss/draw](https://github.com/arcomage/arcomage-hd/issues/84), add "Review cards" button
+- Optimize [AnimatedNumber](https://github.com/arcomage/arcomage-hd/issues/95), Tower/Wall/card anim, & lang change
 - [Debounced number input](https://github.com/arcomage/arcomage-hd/issues/81)
 - [LTR / Arabic support](https://github.com/arcomage/arcomage-hd/issues/90)
 - Fixes and improvements to accessibility (a11y)